refactor: move SpinalSimStreamFifoMultiChannelSharedSpaceTester
