+++ lib/src/test/scala/spinal/lib/Stream.scala
@@ -8,6 +8,7 @@ import spinal.lib.sim._
 
 import spinal.tester.{SpinalTesterCocotbBase, SpinalTesterGhdlBase, SpinalSimFunSuite}
 
+import org.scalatest.funsuite.AnyFunSuite
 import scala.collection.mutable
 import scala.util.Random
 
@@ -2279,6 +2280,40 @@ object TesterBugPlay extends App{
         }
       }
       testbench(dut)
+    }
+  }
+}
+
+class SpinalSimStreamFifoMultiChannelSharedSpaceTester extends AnyFunSuite {
+  test("t1") {
+    SimConfig.compile(new StreamFifoMultiChannelSharedSpace(Bits(32 bits), 4, 16)).doSimUntilVoid(seed = 42) { dut =>
+      val queueModel = mutable.ArrayBuffer.fill(4)(mutable.Queue[Long]())
+
+      SimTimeout(1000000)
+      dut.clockDomain.forkStimulus(2)
+      //Push data randomly and fill the queueModel with pushed transactions
+      dut.io.push.stream.valid #= false
+      dut.io.pop.stream.ready #= true
+
+      val successCount = Array.fill(4)(0)
+      dut.clockDomain.onSamplings {
+        assert(!(dut.io.push.full.toBoolean && (dut.io.availability.toInt > 1)))
+
+        if (dut.io.push.stream.valid.toBoolean && dut.io.push.stream.ready.toBoolean) {
+          queueModel(log2Up(dut.io.push.channel.toInt)).enqueue(dut.io.push.stream.payload.toLong)
+        }
+        if (dut.io.pop.stream.valid.toBoolean && dut.io.pop.stream.ready.toBoolean) {
+          val channel = log2Up(dut.io.pop.channel.toInt)
+          assert(dut.io.pop.stream.payload.toLong == queueModel(channel).dequeue())
+          successCount(channel) += 1
+          if (successCount.forall(_ > 20000)) simSuccess()
+        }
+        dut.io.push.stream.valid.randomize()
+        dut.io.push.stream.payload.randomize()
+        dut.io.push.channel #= (1 << Random.nextInt(dut.channelCount))
+        dut.io.pop.stream.ready.randomize()
+        dut.io.pop.channel #= (1 << Random.nextInt(dut.channelCount))
+      }
     }
   }
 }
